requirements.txt
asgiref==3.5.0
asn1crypto @ file:///tmp/build/80754af9/asn1crypto_1596577642040/work
backports.zoneinfo==0.2.1
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.0
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
defusedxml @ file:///home/ktietz/src/ci_mi/defusedxml_1612808095953/work
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.9.0
django-filter==21.1
django-registration==3.2
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.25.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.0
itypes==1.2.0
Jinja2==3.0.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jwt==1.3.1
kiwisolver @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/kiwisolver_1638569885739/work
mailer @ file:///Users/runner/miniforge3/conda-bld/mailer_1610318509794/work
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib @ file:///opt/concourse/worker/volumes/live/4863dc73-443d-42f7-68bf-489aaa0e23a1/volume/matplotlib-suite_1638289695940/work
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/f196a661-8e33-4aac-63bc-efb2ff50e035/volume/mkl_random_1626186080069/work
mkl-service==2.4.0
munkres==1.1.4
mysqlclient==1.4.6
numexpr @ file:///opt/concourse/worker/volumes/live/52f18ae0-3912-48f4-7832-7d22abf43616/volume/numexpr_1640704220628/work
numpy @ file:///opt/concourse/worker/volumes/live/ceac9fd4-1ae1-48f8-612a-0c99bf783d0e/volume/numpy_and_numpy_base_1634106709647/work
oauthlib==3.1.0
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
Pillow==7.0.0
psycopg2 @ file:///opt/concourse/worker/volumes/live/d12e789f-5b6c-4947-7cc2-187f3a9c9696/volume/psycopg2_1612298599991/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==2.1.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-social-auth==0.3.6
python3-openid @ file:///opt/concourse/worker/volumes/live/f9700ef0-97f0-4528-5d04-8f0c19fb5c75/volume/python3-openid_1614085597480/work
pytz==2019.3
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
scikit-learn @ file:///opt/concourse/worker/volumes/live/09b793c0-41c8-469b-7e31-b4a39bc68d95/volume/scikit-learn_1642617120730/work
scipy @ file:///opt/concourse/worker/volumes/live/7554f9b4-1616-4d3d-50f8-914545975614/volume/scipy_1641555024289/work
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.2
sqlparse==0.3.1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tornado @ file:///opt/concourse/worker/volumes/live/05341796-4198-4ded-4a9a-332fde3cdfd1/volume/tornado_1606942323372/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
uritemplate==4.1.1
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
zipp==3.7.0
